+++ sdk/resourcemanager/appservice/armappservice/containerappsrevisions_client.go
@@ -224,7 +224,6 @@ func (client *ContainerAppsRevisionsClient) getRevisionHandleResponse(resp *http
 }
 
 // NewListRevisionsPager - Get the Revisions for a given Container App.
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-03-01
 // resourceGroupName - Name of the resource group to which the resource belongs.
 // containerAppName - Name of the Container App for which Revisions are needed.