+++ sdk/resourcemanager/appservice/armappservice/workflowrunactions_client.go
@@ -128,7 +128,6 @@ func (client *WorkflowRunActionsClient) getHandleResponse(resp *http.Response) (
 }
 
 // NewListPager - Gets a list of workflow run actions.
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-03-01
 // resourceGroupName - Name of the resource group to which the resource belongs.
 // name - Site name.
@@ -213,7 +212,6 @@ func (client *WorkflowRunActionsClient) listHandleResponse(resp *http.Response) 
 }
 
 // NewListExpressionTracesPager - Lists a workflow run expression trace.
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-03-01
 // resourceGroupName - Name of the resource group to which the resource belongs.
 // name - Site name.