+++ sdk/resourcemanager/appservice/armappservice/workflowtriggers_client.go
@@ -190,7 +190,6 @@ func (client *WorkflowTriggersClient) getSchemaJSONHandleResponse(resp *http.Res
 }
 
 // NewListPager - Gets a list of workflow triggers.
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-03-01
 // resourceGroupName - Name of the resource group to which the resource belongs.
 // name - Site name.